refactor(comments): comments now nested server-side and sent to client ready for rendering
